+++ stac_fastapi/elasticsearch/stac_fastapi/elasticsearch/config.py
@@ -56,6 +56,10 @@ def _es_config() -> Dict[str, Any]:
     if (u := os.getenv("ES_USER")) and (p := os.getenv("ES_PASS")):
         config["http_auth"] = (u, p)
 
+    # Include timeout setting if set
+    if request_timeout := os.getenv("ES_TIMEOUT"):
+        config["request_timeout"] = request_timeout
+
     # Explicitly exclude SSL settings when not using SSL
     if not use_ssl:
         return config