+++ stac_fastapi/opensearch/stac_fastapi/opensearch/config.py
@@ -1,4 +1,5 @@
 """API configuration."""
+
 import logging
 import os
 import ssl
@@ -52,6 +53,10 @@ def _es_config() -> Dict[str, Any]:
             config["headers"] = {"x-api-key": api_key}
 
         config["headers"] = headers
+
+    # Include timeout setting if set
+    if timeout := os.getenv("ES_TIMEOUT"):
+        config["timeout"] = timeout
 
     # Explicitly exclude SSL settings when not using SSL
     if not use_ssl: